Change PluginManager::pluginCount() to return size_t
This changes the PluginManager::pluginCount() method
to return a size_t instead of an int.

+++ src/chemkit/pluginmanager.cpp
@@ -106,7 +106,7 @@ const std::vector<Plugin *>& PluginManager::plugins() const
 }
 
 /// Returns the number of loaded plugins.
-int PluginManager::pluginCount() const
+size_t PluginManager::pluginCount() const
 {
     return d->plugins.size();
 }
